Update BLEScalePlugin.cpp for Aliexpress myscale support (#353)
* Update BLEScalePlugin.cpp for Aliexpress myscale support

Added call to myscale Aliexprss cheap scale in conjunction with https://github.com/gaggimate/esp-arduino-ble-scales/pull/11

* Fixed issue where the bluetooth scale was taring too late (at brew start).
If the scale had a higher value than the voulumetric setpoint at start,
the tare would not happen in time and the brew would finish.

The controller now tares the bluetooth scale before starting the brew and at the same time as it issues controller->tare() to the hardware scale.